Re-Organised Dashboard Scripts and pages
Added partial html pages for channel overview and messages.
Split js out into appropriate files.
