+++ Historian/HIstorian.Service/Historian.Service.csproj
@@ -120,6 +120,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Controllers\DashboardChannelsController.cs" />
     <Compile Include="Controllers\MessagesController.cs" />
     <Compile Include="Controllers\ChannelsController.cs" />
     <Compile Include="HistorianService.cs">
