+++ Historian/Historian.Dashboard/Dashboard/Content/html/dashboard.channel.messages.html
@@ -1,11 +1,22 @@
-﻿<!DOCTYPE html>
-
-<html lang="en" xmlns="http://www.w3.org/1999/xhtml">
-<head>
-    <meta charset="utf-8" />
-    <title></title>
-</head>
-<body>
-
-</body>
-</html>
+﻿<h3 class="page-header">
+    Messages
+    <small class="historian-dashboard-title"></small>
+</h3>
+<div class="table-responsive">
+    <table class="table table-striped historian-messages">
+        <thead>
+            <tr>
+                <th>Timestamp</th>
+                <th>Title</th>
+                <th>Tags</th>
+                <th>Kind</th>
+                <th>Contents</th>
+            </tr>
+        </thead>
+        <tbody>
+            <tr>
+                <td colspan="6">You have no messages</td>
+            </tr>
+        </tbody>
+    </table>
+</div>
