+++ Historian/Historian.Dashboard/Dashboard/Content/html/dashboard.html
@@ -12,7 +12,6 @@
 
     <!-- Bootstrap core CSS -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css" integrity="sha384-1q8mTJOASx8j1Au+a5WDVnPi2lkFfwwEAa8hDDdjZlpLegxhjVME1fgjWPGmkzs7" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap-theme.min.css" integrity="sha384-fLW2N01lMqjakBkx3l/M9EahuwpSfeNvV63J5ezn3uZzapT0u7EYsXMjQV+0En5r" crossorigin="anonymous">
 
     <!-- Custom styles for this template -->
     <link href="{baseUrl}/content/css/dashboard.css" rel="stylesheet">
@@ -73,35 +72,14 @@
             </div>
             <div class="col-sm-9 col-sm-offset-3 col-md-10 col-md-offset-2 main">
                 <h1 class="page-header">
-                    Dashboard
+                    Overview
                     <small class="historian-dashboard-title"></small>
                 </h1>
                 <div class="row placeholders">
                     <div class="col-xs-6 col-sm-3 placeholder">
-                        <canvas class="historian-messages-kindsChart" width="200" height="200"></canvas>
+                        <canvas class="historian-messages-kindsChart" width="400" height="400"></canvas>
                         <h4>Distribution by Kind</h4>
                     </div>
-                </div>
-
-                <h2 class="sub-header">Messages</h2>
-                <div class="table-responsive">
-                    <table class="table table-striped historian-messages">
-                        <thead>
-                            <tr>
-                                <th>#</th>
-                                <th>Title</th>
-                                <th>Tags</th>
-                                <th>Kind</th>
-                                <th>Contents</th>
-                                <th>Timestamp</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr>
-                                <td colspan="6">You have no messages</td>
-                            </tr>
-                        </tbody>
-                    </table>
                 </div>
             </div>
         </div>
@@ -114,9 +92,13 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js" integrity="sha384-0mSbJDEHialfmuBBQP6A4Qrprq5OVfW37PRR3j5ELqxss1yVqOtnepnHVP9aJ7xS" crossorigin="anonymous"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.0.1/Chart.min.js"></script>
     <script src="{baseUrl}/content/js/dashboard.js"></script>
+    <script src="{baseUrl}/content/js/dashboard.sidebar.js"></script>
+    <script src="{baseUrl}/content/js/dashboard.channel.overview.js"></script>
+    <script src="{baseUrl}/content/js/dashboard.channel.messages.js"></script>
     <script>
-        var dashboard = new hsDashboard();
-        dashboard.load();
+        $(function () {
+            historian.load();
+        });
     </script>
 </body>
 </html>
