+++ Historian/Historian.Dashboard/Dashboard/Content/html/dashboard.sidebar.html
@@ -1,0 +1,11 @@
+﻿<ul class="nav nav-sidebar">
+    <li>
+        <a href="#">
+            <span class="glyphicon glyphicon-stats" aria-hidden="true"></span>
+            Overview
+        </a>
+    </li>
+</ul>
+<ul class="nav nav-sidebar historian-sidebar-channels">
+    <li><a href="#">Loading...</a></li>
+</ul>
